[CI Visibility] Refactor gotesting integration to be used by orchestrion
